Created Queue Application.
The application simply creates a list of things.

The main thing I need to redo is the post via Thymeleaf. Not that's bad, it's just that I couldn't figure out hot to properly implement the post via Vue.js and Axios.
There are a lot of other TODOs, but everything is functional at this point for the simple purpose intended. Anyway, here's a list of TODOs:
- Properly return date or datetime and process it on the frontend.
- Use Command Objects (DTOs) instead of inner classes and transformations on the Controllers.
- Order and Paginate the repository and service results.
- Configure a proper database instead of H2.